accessors and access modifiers overhaul
