+++ NccsvConverter/NccsvConverter.ConsoleApp/ConsoleApp.cs
@@ -2,7 +2,7 @@
 using NccsvConverter.NccsvParser.Repositories;
 using NccsvConverter.NccsvSerializer.SchemaDatasetJsonSerializer;
 
-Console.WriteLine("Write name of file or press enter to quit");
+Console.WriteLine("If you want to validate, write name of file:");
 string filePathFolder = Directory.GetParent(Directory.GetCurrentDirectory()).Parent.Parent.Parent.FullName
                         + "\\NccsvConverter.ConsoleApp\\TestData\\";
 string fileName = Console.ReadLine();
@@ -23,17 +23,12 @@ if (File.Exists(filePath))
     {
         Console.WriteLine("No problems found.");
 
-        // TODO: make optional
-        Serializer.ToJson(dataSet);
+        Console.WriteLine("If you want to serialize, press y:");
+        var input = Console.ReadKey();
+        Console.WriteLine();
+        if (input.KeyChar == 'y')
+            Console.WriteLine(Serializer.ToJson(dataSet));
     }
 }
 else
     Console.WriteLine($"File \"{filePath}\" could not be found.");
-
-
-
-
-
-
-
-
