+++ NccsvConverter/NccsvConverter.NccsvParser/Models/DataSet.cs
@@ -1,20 +1,21 @@
 ﻿using NccsvConverter.NccsvParser.Helpers;
 using NccsvConverter.NccsvParser.Repositories;
-using System.Data;
 using NccsvConverter.NccsvParser.Validators;
+using System.Globalization;
 
 namespace NccsvConverter.NccsvParser.Models;
 
 public class DataSet
 {
-    private MetaData _metaData = new MetaData();
-    public MetaData MetaData { get; set; }
-    public List<DataValue[]> Data { get; set; } = new();
-
-
-    public DataSet()
+    private MetaData _metaData; 
+    public MetaData MetaData
     {
-        MetaData = _metaData;
+        get { return _metaData; }
+    }
+    private List<DataValue[]> _data = new();
+    public List<DataValue[]> Data 
+    {
+        get { return _data; }
     }
 
 
@@ -64,14 +65,14 @@ public class DataSet
                     if (line == "*END_DATA*")
                     {
                         endDataFound = true;
-                        DataValidator.Validate(endDataFound);
+                        DataValidator.Validate(endDataFound, headers);
                         break;
                     }
                     else if (headersFound)
                     {
                         //TODO: verify headers? check for scalar variables
                         MetaDataValidator.Validate(metaDataList, endMetaDataFound);
-                        dataSet.MetaDataHandler(metaDataList);
+                        dataSet._metaData = new MetaData(metaDataList);
                         headers = separatedLine;
                         headersFound = false;
                         continue;
@@ -92,45 +93,323 @@ public class DataSet
     }
 
 
-    private void MetaDataHandler(List<string[]> metaDataList)
-    {
-        // Find and add global attributes to dataset
-        var globalAttributes = NccsvParserMethods.FindGlobalAttributes(metaDataList);
-
-        MetaData.Title = NccsvParserMethods.FindTitle(globalAttributes);
-        MetaData.Summary = NccsvParserMethods.FindSummary(globalAttributes);
-
-        NccsvParserMethods.AddGlobalAttributes(this, globalAttributes);
-
-        // Find variable metadata
-        var variableMetaData = NccsvParserMethods.FindVariableMetaData(metaDataList);
-
-        if (!VariableMetaDataValidator.Validate(variableMetaData))
-            return;
-
-        // Create variables from variable metadata and add to dataset
-        foreach (var line in variableMetaData)
-        {
-            if (!NccsvParserMethods.CheckIfVariableExists(MetaData.Variables, line[0]))
-            {
-                var varToCreate = NccsvParserMethods.IsolateVariableAttributes(variableMetaData, line[0]);
-
-                var newVariable = NccsvParserMethods.CreateVariable(varToCreate);
-
-                NccsvParserMethods.SetVariableDataType(newVariable, varToCreate);
-
-                //TODO: Verify variable?
-
-                MetaData.Variables.Add(newVariable);
-            }
-        }
-    }
-
-
     private void DataRowHandler(string[] dataRow, string[] headers, int rowNumber, bool saveData)
     {
         if (!DataRowValidator.Validate(dataRow, headers, rowNumber))
             return;
-        NccsvParserMethods.AddData(dataRow, headers, this, rowNumber, saveData);
+        SetData(dataRow, headers, rowNumber, saveData);
+    }
+
+
+    private void SetData(string[] dataRow, string[] headers, int rowNumber, bool saveData)
+    {
+        List<DataValue> dataValues = new();
+
+        for (int i = 0; i < dataRow.Length; i++)
+        {
+            var variable = _metaData.Variables
+                .FirstOrDefault(v => v.VariableName
+                    .Equals(headers[i]));
+
+            if (variable != null)
+            {
+                var dataValue = CreateDataValueAccordingToDataType(dataRow[i], variable);
+
+                if (dataValue != null)
+                {
+                    if (saveData)
+                        dataValues.Add(dataValue);
+                }
+                else
+                {
+                    MessageRepository.Messages.Add(
+                        new Message($"Row {rowNumber}: Data value \"{dataRow[i]}\" could not be parsed to variable datatype: {variable.VariableDataType}.", Severity.NonCritical));
+                }
+            }
+            else
+            {
+                MessageRepository.Messages.Add(
+                    new Message($"Header \"{dataRow[i]}\" did not match any variables.", Severity.NonCritical));
+            }
+        }
+
+        if (saveData)
+            _data.Add(dataValues.ToArray());
+    }
+
+
+    // Creates and returns a DataValueAs<T> from a given value and variable,
+    // where T is the DataType of the variable that acts as column header.
+    // If unsuccessfull, returns a null value.
+    private static DataValue? CreateDataValueAccordingToDataType(string value, Variable variable)
+    {
+        bool result;
+
+        switch (variable.VariableDataType)
+        {
+            case "byte":
+                // byte -> c# sbyte
+
+                if (value != "")
+                { 
+                    result = sbyte.TryParse(value, out sbyte byteValue);
+
+                    if (result)
+                        return new DataValueAs<sbyte>()
+                        {
+                            Variable = variable,
+                            Value = byteValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<sbyte?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+
+                return null;
+
+            case "ubyte":
+                // unsigned byte -> c# byte
+
+                if (value != "")
+                {
+                    result = byte.TryParse(value, out byte ubyteValue);
+
+                    if (result)
+                    return new DataValueAs<byte>()
+                    {
+                        Variable = variable,
+                        Value = ubyteValue
+                    };
+                }
+                else
+                {
+                    return new DataValueAs<byte?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+                
+                return null;
+
+            case "short":
+
+                if (value != "")
+                {
+                     result = short.TryParse(value, out short shortValue);
+
+                    if (result)
+                        return new DataValueAs<short>()
+                        {
+                            Variable = variable,
+                            Value = shortValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<short?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+
+                return null;
+
+            case "ushort":
+
+                if (value != "")
+                {
+                    result = ushort.TryParse(value, out ushort ushortValue);
+
+                    if (result)
+                        return new DataValueAs<ushort>()
+                        {
+                            Variable = variable,
+                            Value = ushortValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<ushort?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+                
+                return null;
+
+            case "int":
+
+                if (value != "")
+                {
+                    result = int.TryParse(value, out int intValue);
+
+                    if (result)
+                        return new DataValueAs<int>()
+                        {
+                            Variable = variable,
+                            Value = intValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<int?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+                
+                return null;
+
+            case "uint":
+
+                if (value != "")
+                {
+                    result = uint.TryParse(value, out uint uintValue);
+
+                    if (result)
+                        return new DataValueAs<uint>()
+                        {
+                            Variable = variable,
+                            Value = uintValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<uint?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+
+                return null;
+
+            case "long":
+                // TODO: check L
+                if (value != "")
+                {
+                    result = long.TryParse(value[..^1], out long longValue);
+
+                    if (result)
+                        return new DataValueAs<long>()
+                        {
+                            Variable = variable,
+                            Value = longValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<long?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+                
+                return null;
+
+            case "ulong":
+                // TODO: check uL
+                if (value != "")
+                {
+                    result = ulong.TryParse(value[..^2], out ulong ulongValue);
+
+                    if (result)
+                        return new DataValueAs<ulong>()
+                        {
+                            Variable = variable,
+                            Value = ulongValue
+                        };
+                }
+                else
+                {
+                    return new DataValueAs<ulong?>()
+                    {
+                        Variable = variable,
+                        Value = null
+                    };
+                }
+                return null;
+
+            case "float":
+
+                if (value != "")
+                {
+                    result = float.TryParse(value, CultureInfo.InvariantCulture, out float floatValue);
+
+                    if (result)
+                    return new DataValueAs<float>()
+                    {
+                        Variable = variable,
+                        Value = floatValue
+                    };
+                }
+                else
+                {
+                    return new DataValueAs<float>()
+                    {
+                        Variable = variable,
+                        Value = float.NaN
+                    };
+                }
+
+                return null;
+
+            case "double":
+
+                if (value != "")
+                {
+                    result = double.TryParse(value, CultureInfo.InvariantCulture, out double doubleValue);
+
+                    if (result)
+                    return new DataValueAs<double>()
+                    {
+                        Variable = variable,
+                        Value = doubleValue
+                    };
+                }   
+                else
+                {
+                    return new DataValueAs<double>()
+                    {
+                        Variable = variable,
+                        Value = double.NaN
+                    };
+                }
+                
+                return null;
+
+            case "string":
+                return new DataValueAs<string>
+                {
+                    Variable = variable,
+                    Value = value
+                };
+
+            case "char":
+                // TODO: handle special char cases
+                result = char.TryParse(value, out char charValue);
+
+                if (result)
+                    return new DataValueAs<char>()
+                    {
+                        Variable = variable,
+                        Value = charValue
+                    };
+                else
+                    return null;
+
+            default:
+                return null;
+        }
     }
 }
