+++ NccsvConverter/NccsvConverter.NccsvParser/Validators/VariableMetaDataValidator.cs
@@ -1,6 +1,4 @@
-﻿
-using NccsvConverter.NccsvParser.Helpers;
-using NccsvConverter.NccsvParser.Models;
+﻿using NccsvConverter.NccsvParser.Models;
 using NccsvConverter.NccsvParser.Repositories;
 
 namespace NccsvConverter.NccsvParser.Validators;
