+++ NccsvConverter/NccsvConverter.TestProject/Usings.cs
@@ -1,4 +1,3 @@
 global using Xunit;
-global using NccsvConverter.NccsvParser.FileHandling;
 global using NccsvConverter.NccsvParser.Helpers;
